Bring more structure into tests
